Added the nTsMapper nuget package to the project which scans your C# code to find domain objects and DTOs and generates TypeScript interfaces and classes so that the client side TypeScript code can use those classes and interfaces instead of any.
- updated the default Json serialization behavior to camelCase the property names so that they match the JavaScript standard
- updated SearchArea/SearchAreaVm.ts and SearchArea/SearchAreaVmSpecs.ts to use concrete types instead of any

This code is currently not stable, because nTsMapper currently tries to map byte arrays as a JavaScript collection of numbers (which you would expect), except that the default JSON serialization behavior is to serialize byte arrays as base64 encoded strings.  However, I altered the PeopleSearch.Api/PeopleSearch.Api.ts file to compensate for this behavior, but it will be broken on the next rebuild.